Retrieve country name from the content
The length of sheet name seems to have limit of 31 characters on length.
That causes the names of the following countries to be truncated,
French Southern and Antarctic Lands turns to be
French Southern and Antarctic L.

Democratic Republic of the Congo turns to be
Democratic Republic of the Cong.

This change get the intact country name from the content instead of
from the sheet name.