Update or create overview item
Check if the overview item with specific date and region already exists
or not.
If it already exists, update the new avg and max instead of creating a
new entry.

Add "MultipleActiveResultSets=true" to the provider part of connection
string to solve the following exception,
Unhandled exception. System.InvalidOperationException: There is already
an open DataReader associated with this Command which must be closed
first.

Refer to the following stackoverflow ticket for more details.
https://stackoverflow.com/a/6064422/8735281