Avoid blocking import due to file move exception
If we'd like to move an imported file to 'Processed' directory but a
file with the same file name has already existed, the following
exception would be raised and block the following file handlings.

System.IO.IOException: Cannot create a file when that file already exists.

This commit logs the error and keeps importing other files to Db
instead of blocking the rest of operation.