Make some fields read-only
Capitalize the naming of content and reportDate variables to abide by
property convention.

Remove the setter from Content and ReportDate properties and initialize
them in the constructor.

Change method name from ImportTo to ExportTo to improve readibility.

Change modifier from virtual to abstract because the implementations
are only needed in derived classes accordingly.