Ignore the game detail with invalid release date
The "Party Animals" game in "GameRanks_20201007.xlsx" has the
"Release Date" of "Late 2020" and caused the following exception
because of invalid DateTime format.

System.FormatException: The string 'Late 2020' was not recognized as a
valid DateTime. There is an unknown word starting at index '0'.

This commit ignores the game and bypasses the exception.